requirements.txt
streamlit==1.54.0
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
easyocr==1.7.2
numpy==1.26.4
pillow==12.1.1
ultralytics==8.4.15
pandas==2.3.3
torch==2.0.1